refactor: moves handlerbars setup into separate file for re-use by test renderer, fixes template paths
